Edição e exclusão de forncedores

+++ CrudVega/Repositories/SupplierRepository.cs
@@ -1,6 +1,7 @@
 ﻿using CrudVega.Context;
 using CrudVega.Models;
 using Microsoft.AspNetCore.Http.HttpResults;
+using Microsoft.AspNetCore.Mvc;
 
 namespace CrudVega.Repositories
 {
@@ -18,15 +19,39 @@ namespace CrudVega.Repositories
             return supplier;
         }
 
-        public void DeleteSupplier(SupplierModel supplier)
+        public void DeleteSupplier(int id)
         {
-            _context.Suppliers.Remove(supplier);
+            var supplier = GetSupplier(id);
+            if (supplier != null)
+            {
+                //Tratamento de exceção aqui para exclusão de fornecedor. Tem produto vinculado ao fornecedor e esse motivo não o deixa excluir o fornecedor
+                _context.Suppliers.Remove(supplier);
+                _context.SaveChanges();
+            }
+        }
+
+        public void EditSupplier(SupplierModel supplier)
+        {
+            _context.Suppliers.Update(supplier);
             _context.SaveChanges();
         }
 
         public IEnumerable<SupplierModel> GetAllSuppliers()
         {
             return _context.Suppliers.ToList();
+        }
+
+        public SupplierModel GetSupplier(int id)
+        {
+            var supplier = _context.Suppliers.FirstOrDefault(s => s.Id == id);
+            if (supplier == null)
+            {
+                return null;
+            }
+            else
+            {
+                return supplier;
+            }
         }
     }
 }
